refactor(evals): extract synchronized scroll logic into custom hook
- Create useSynchronizedScroll hook in hooks/useSynchronizedScroll.ts
- Simplifies RemapEvalWizard by removing inline useEffect
- Reusable hook for any dual-panel synchronized scrolling
- Improves code organization and testability
